Create a config struct and pass it in userdata.
So far we only pass a single parameter–the verbosity–in userdata, so we got away with passing a bare int. However it is a bit confusing.  Wrapping it in a struct should be easier to understand what's going on. Also the struct is more flexible in case we want to pass more
parameters.